requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
chardet==4.0.0
frozenlist==1.5.0
idna==2.10
multidict==6.1.0
paho-mqtt==2.1.0
propcache==0.2.0
PyJWT==2.10.1
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
schedule==1.2.2
six==1.16.0
twilio==9.3.7
typing_extensions==4.12.2
urllib3==1.26.20
websocket-client==1.8.0
yarl==1.18.0
